+++ oca/utils/__init__.py
@@ -5,6 +5,7 @@ Utilities
 
 from .cli import *
 from .config import *
+from .constants import *
 from .experience_replay import *
 from .logger import *
 from .option_critic import *